+++ x/distribution/keeper/delegation.go
@@ -248,7 +248,7 @@ func (k Keeper) withdrawDelegationRewards(ctx context.Context, val stakingtypes.
 		)
 	}
 
-	// truncate reward dec coins, return remainder to community pool
+	// truncate reward dec coins, return remainder to decimal pool
 	finalRewards, remainder := rewards.TruncateDecimal()
 
 	// add coins to user account
@@ -264,10 +264,8 @@ func (k Keeper) withdrawDelegationRewards(ctx context.Context, val stakingtypes.
 		}
 	}
 
-	// update the outstanding rewards and the community pool only if the
-	// transaction was successful
-	err = k.ValidatorOutstandingRewards.Set(ctx, sdk.ValAddress(valAddr), types.ValidatorOutstandingRewards{Rewards: outstanding.Sub(rewards)})
-	if err != nil {
+	// update the outstanding rewards and the decimal pool only if the transaction was successful
+	if err := k.ValidatorOutstandingRewards.Set(ctx, sdk.ValAddress(valAddr), types.ValidatorOutstandingRewards{Rewards: outstanding.Sub(rewards)}); err != nil {
 		return nil, err
 	}
 
@@ -276,7 +274,7 @@ func (k Keeper) withdrawDelegationRewards(ctx context.Context, val stakingtypes.
 		return nil, err
 	}
 
-	feePool.CommunityPool = feePool.CommunityPool.Add(remainder...)
+	feePool.DecimalPool = feePool.DecimalPool.Add(remainder...)
 	err = k.FeePool.Set(ctx, feePool)
 	if err != nil {
 		return nil, err